Refactor library build task to use copy-files module
