+++ lib/tasks/build.js
@@ -6,25 +6,30 @@ var path = require('path');
 
 var ui = require('../ui');
 var Promise = require('../ext/promise');
+var copyFiles = require('../utils/copy-files');
 
 var copy = Promise.denodeify(fs.copy);
 var readFile = Promise.denodeify(fs.readFile);
 
-function determineMicroAddonType(packageJsonContent) {
-  var packageInfo = JSON.parse(packageJsonContent);
-  var addonType;
+// common
 
-  if (packageInfo.keywords) {
-    if (packageInfo.keywords.indexOf('ember-micro-component') >= 0) {
-      addonType = 'component';
-    } else if (packageInfo.keywords.indexOf('ember-micro-helper') >= 0) {
-      addonType = 'helper';
-    } else if (packageInfo.keywords.indexOf('ember-micro-library') >= 0) {
-      addonType = 'library';
+function determineMicroAddonType(microAddon) {
+  return readFile(path.join(microAddon.path, 'package.json')).then(function(fileContent) {
+    var packageInfo = JSON.parse(fileContent);
+    var addonType;
+
+    if (packageInfo.keywords) {
+      if (packageInfo.keywords.indexOf('ember-micro-component') >= 0) {
+        addonType = 'component';
+      } else if (packageInfo.keywords.indexOf('ember-micro-helper') >= 0) {
+        addonType = 'helper';
+      } else if (packageInfo.keywords.indexOf('ember-micro-library') >= 0) {
+        addonType = 'library';
+      }
     }
-  }
 
-  return addonType;
+    return addonType;
+  });
 }
 
 function writeIndexJs(microAddon) {
@@ -37,7 +42,7 @@ function writeIndexJs(microAddon) {
   indexFile.write('};\n');
   indexFile.end();
 
-  return Promise.resolve(microAddon);
+  return Promise.resolve();
 }
 
 function copyPackageJson(microAddon) {
@@ -52,14 +57,6 @@ function copyHelperJs(microAddon) {
   var oldHelper = path.join(microAddon.path, 'helper.js');
   var newHelper = path.join(microAddon.path, 'dist', 'app', 'helpers', microAddon.name + '.js');
   return copy(oldHelper, newHelper, { clobber: true }).then(function() {
-    return Promise.resolve(microAddon);
-  });
-}
-
-function copyLibraryJs(microAddon) {
-  var oldLibrary = path.join(microAddon.path, 'library.js');
-  var newLibrary = path.join(microAddon.path, 'dist', 'app', 'lib', microAddon.name + '.js');
-  return copy(oldLibrary, newLibrary, { clobber: true }).then(function() {
     return Promise.resolve(microAddon);
   });
 }
@@ -88,6 +85,8 @@ function copyComponentStyle(microAddon) {
   });
 }
 
+// helper
+
 function buildMicroHelper(microAddon) {
   ui.write('Packaging helper...', microAddon.name);
   return copyPackageJson(microAddon)
@@ -95,12 +94,30 @@ function buildMicroHelper(microAddon) {
     .then(copyHelperJs);
 }
 
+// library
+
+function getLibraryMappings(microAddon) {
+  return [
+    {
+      oldPath: path.join(microAddon.path, 'package.json'),
+      newPath: path.join(microAddon.path, 'dist', 'package.json')
+    },
+    {
+      oldPath: path.join(microAddon.path, 'library.js'),
+      newPath: path.join(microAddon.path, 'dist', 'app', 'lib', microAddon.name + '.js')
+    }
+  ];
+}
+
 function buildMicroLibrary(microAddon) {
   ui.write('Packaging library...', microAddon.name);
-  return copyPackageJson(microAddon)
-    .then(writeIndexJs)
-    .then(copyLibraryJs);
+  var fileMappings = getLibraryMappings(microAddon);
+  return copyFiles(fileMappings).then(function() {
+    return writeIndexJs(microAddon);
+  });
 }
+
+// component
 
 function buildMicroComponent(microAddon) {
   ui.write('Packaging component...', microAddon.name);
@@ -121,12 +138,11 @@ function buildMicroAddon(microAddon, type) {
   }
 }
 
+
 module.exports = function(options) {
   var microAddon = { name: options.name, path: path.join(process.cwd(), options.name) };
 
-  return readFile(path.join(microAddon.path, 'package.json'))
-    .then(determineMicroAddonType)
-    .then(function(type) {
-      return buildMicroAddon(microAddon, type);
-  	});
+  return determineMicroAddonType(microAddon).then(function(type) {
+    return buildMicroAddon(microAddon, type);
+	});
 };
