+++ lib/utils/copy-files.js
