+++ lib/utils/extract-addon.js
@@ -4,7 +4,7 @@ var fs = require('fs-extra');
 var path = require('path');
 
 var Promise = require('../ext/promise');
-var overwriteFiles = require('./overwrite-files');
+var copyFiles = require('./copy-files');
 
 var ensureDir = Promise.denodeify(fs.ensureDir);
 var copy = Promise.denodeify(fs.copy);
@@ -44,6 +44,6 @@ module.exports = function(options) {
   }).then(function() {
     return replaceInFiles(options.filesToReplacePlaceholderIn, options.placeholder, options.addonName);
   }).then(function () {
-    return overwriteFiles(options.fileMappings);
+    return copyFiles(options.fileMappings);
   });
 };
